Переносы кода в Kernel и Scene

+++ Novel Fox Engine/Engine/sprite.cpp
@@ -86,8 +86,8 @@ void Sprite::computeLayerScale()
 	int h = this->getTextureRect().height;
 	// ВНИМАНИЕ: константа выведена эмпирически, значит это лишь частный случай
 	// Рассчитываем scale в зависимости от слоя (квадратичная зависимость)
-	float sx = posScale.scale.x + 0.03 * (2 << (layer-1));
-	float sy = posScale.scale.y + 0.03 * (2 << (layer-1));
+	float sx = posScale.scale.x + (float)0.03 * (2 << (layer-1));
+	float sy = posScale.scale.y + (float)0.03 * (2 << (layer-1));
 	// Теперь выставляем новые координаты за вычетом увеличения пополам
 	posScale.pos.x = posScale.pos.x - w * (sx - posScale.scale.x) / 2;
 	posScale.pos.y = posScale.pos.y - h * (sy - posScale.scale.y) / 2;
